creating upgrade file and upgrading zap files when opened
