+++ src-electron/db/query-cluster.js
@@ -23,15 +23,6 @@
 const dbApi = require('./db-api.js')
 const dbMapping = require('./db-mapping.js')
 
-async function selectClusterName(db, clusterRef) {
-  let clusterName = await dbApi.dbAll(
-    db,
-    'SELECT NAME FROM CLUSTER WHERE CLUSTER_ID = ?',
-    [clusterRef]
-  )
-  return clusterName[0].NAME
-}
-
 /**
  * All cluster details along with their attribute details per endpoint.
  * @param db
@@ -288,7 +279,6 @@ async function selectTokenAttributeClustersForEndpoint(
   )
   return rows.map(dbMapping.map.cluster)
 }
-exports.selectClusterName = selectClusterName
 exports.selectClusterDetailsFromEnabledClusters =
   selectClusterDetailsFromEnabledClusters
 exports.selectAllUserClustersWithTokenAttributes =
